+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
@@ -6,29 +7,45 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Migrate to typescript.
+- Change class name `ElementSort` -> `ElementSorter`.
+
 ## [1.1.3] - 2017-12-17
+
 ### Fixed
+
 - Fix problem in `findAncestor()`. #2
 
 ## [1.1.2] - 2017-12-11
+
 ### Added
+
 - Adds notice about using polyfill for browser compatibility. Please see [README.md#polyfill](https://github.com/archco/element-util/blob/master/README.md#polyfill)
 
 ### Fixed
+
 - Fix compatibility problem in `show()`.
 
 ## [1.1.1] - 2017-11-19
+
 ### Changed
+
 - `findAncestor()`: It works too if set ancestor parameter as Element.
 
 ## [1.1.0] - 2017-10-26
+
 ### Added
+
 - Add util methods: `makeHiddenInput()` and `appendHiddenInput()`.
 
 ### Fixed
+
 - Fix util method `toggleShow()`.
 
 ## [1.0.0] - 2017-10-07
+
 First release. Please see [ElementUtil API](https://github.com/archco/element-util/tree/master/docs).
 
 [Unreleased]: https://github.com/archco/element-util/compare/v1.1.3...HEAD